+++ src/dfx/src/lib/wasm/metadata.rs
@@ -8,15 +8,12 @@ use std::path::Path;
 
 #[context("Failed to add candid service metadata from {} to {}.", idl_path.to_string_lossy(), wasm_path.to_string_lossy())]
 pub fn add_candid_service_metadata(wasm_path: &Path, idl_path: &Path) -> DfxResult {
-    let mut m = walrus::ModuleConfig::new()
-        .parse_file(wasm_path)
-        .with_context(|| format!("Failed to parse file {}", wasm_path.to_string_lossy()))?;
+    let wasm = std::fs::read(wasm_path).context("Could not read the WASM module.")?;
     let idl = std::fs::read(&idl_path)
         .with_context(|| format!("Failed to read {}", idl_path.to_string_lossy()))?;
-
-    add_metadata(&mut m, Kind::Public, CANDID_SERVICE, idl);
-
-    m.emit_wasm_file(wasm_path)
-        .with_context(|| format!("Failed to emit wasm to {}", wasm_path.to_string_lossy()))?;
+    let processed_wasm = add_metadata(&wasm, Kind::Public, CANDID_SERVICE, idl)
+        .context("Could not add metadata to the WASM module.")?;
+    std::fs::write(wasm_path, &processed_wasm)
+        .with_context(|| format!("Could not write WASM to {:?}", wasm_path))?;
     Ok(())
 }